refactor: 문자열 덧셈 계산기 리팩토링
