+++ src/main/java/step1/InValidNumberException.java
@@ -1,0 +1,7 @@
+package step1;
+
+public class InValidNumberException extends RuntimeException {
+    public InValidNumberException(String message) {
+        super(message);
+    }
+}
